# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API de cuentas y movimientos
  description: "Prueba tecnica Java Senior de Julio Flores. Permite crear cuentas,\
    \ consultar saldos y registrar creditos y debitos con idempotencia por cuenta."
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: Generated server url
tags:
- name: Cuentas
  description: Creacion y consulta de cuentas
- name: Actuator
  description: Monitor and interact
  externalDocs:
    description: Spring Boot Actuator Web API Documentation
    url: https://docs.spring.io/spring-boot/docs/current/actuator-api/html/
paths:
  /accounts:
    post:
      tags:
      - Cuentas
      summary: Crear una cuenta con saldo inicial
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
        required: true
      responses:
        "201":
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          description: Saldo inicial o JSON invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /accounts/{id}/transactions:
    post:
      tags:
      - Movimientos
      summary: Registrar un credito o debito
      description: La clave es unica por cuenta. Repetir la misma clave con el mismo
        tipo y monto devuelve el movimiento original sin volver a modificar el saldo.
      operationId: execute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        description: "Clave de 1 a 100 caracteres. Permite letras, numeros, punto,\
          \ guion, guion bajo y dos puntos."
        required: true
        schema:
          type: string
        example: credito-001
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
        required: true
      responses:
        "201":
          description: Movimiento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "200":
          description: Solicitud repetida; devuelve el movimiento original
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Datos invalidos o clave ausente o invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Cuenta inexistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: "Fondos insuficientes, limite de saldo o clave utilizada con\
            \ otros datos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /actuator:
    get:
      tags:
      - Actuator
      summary: Actuator root web endpoint
      operationId: links
      responses:
        "200":
          description: OK
          content:
            application/vnd.spring-boot.actuator.v3+json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties:
                    $ref: "#/components/schemas/Link"
            application/vnd.spring-boot.actuator.v2+json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties:
                    $ref: "#/components/schemas/Link"
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties:
                    $ref: "#/components/schemas/Link"
  /actuator/health:
    get:
      tags:
      - Actuator
      summary: Actuator web endpoint 'health'
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/vnd.spring-boot.actuator.v3+json:
              schema:
                type: object
            application/vnd.spring-boot.actuator.v2+json:
              schema:
                type: object
            application/json:
              schema:
                type: object
  /accounts/{id}:
    get:
      tags:
      - Cuentas
      summary: Consultar una cuenta y su saldo
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Cuenta encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          description: Identificador invalido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Cuenta inexistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    CreateAccountRequest:
      type: object
      properties:
        initialBalance:
          type: number
          minimum: 0.00
      required:
      - initialBalance
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        initialBalance:
          type: number
        balance:
          type: number
    ApiError:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
    CreateTransactionRequest:
      type: object
      properties:
        type:
          type: string
          enum:
          - CREDIT
          - DEBIT
        amount:
          type: number
          minimum: 0.01
      required:
      - amount
      - type
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountId:
          type: string
          format: uuid
        type:
          type: string
          enum:
          - CREDIT
          - DEBIT
        amount:
          type: number
        resultingBalance:
          type: number
        createdAt:
          type: string
          format: date-time
    Link:
      type: object
      properties:
        href:
          type: string
        templated:
          type: boolean